refactor : category 디렉토리 따로 생성 Issue #7
