+++ src/main/java/com/kdt/localinfo/category/Category.java
@@ -1,4 +1,4 @@
-package com.kdt.localinfo.post.Entity;
+package com.kdt.localinfo.category;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
