+++ src/main/java/com/kdt/localinfo/category/CategoryRepository.java
@@ -1,6 +1,5 @@
-package com.kdt.localinfo.post.repository;
+package com.kdt.localinfo.category;
 
-import com.kdt.localinfo.post.Entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
