+++ src/test/java/com/kdt/localinfo/post/controller/PostControllerTest.java
@@ -1,16 +1,15 @@
 package com.kdt.localinfo.post.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.kdt.localinfo.post.Entity.Category;
+import com.kdt.localinfo.category.Category;
 import com.kdt.localinfo.post.dto.PostDto;
-import com.kdt.localinfo.post.repository.CategoryRepository;
+import com.kdt.localinfo.category.CategoryRepository;
 import com.kdt.localinfo.post.service.PostService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
